Initial Commit -
Have created the Git repository and a main branch.

In this initial commit, have added a few lines of code to ensure the
program is working as intended.

JFrame has been imported to build the window that the game will 'sit'
inside of.

Have created a few variable outlining a basic window - these can be
changed at a later date.

Next Task:

Is to finalise the dimensions of the window and create the separate
classes for the game parameters