Second Commit -
Implemented the PongPanel class -

This will eventually house the playing surface of the game.

Used the final variable to set the background color to black.

Added several variables related to the action events related to keys
being pressed.

Added a numbe of comments to explain some of the changes to the project

Pong Class -

added a call to the PongPanel class inside the Pong constructor - as
well as a call to run the panel inside the window