Experiment_2 Branch commit
Sprite class completed -

Nature of the class is to collate all the necessary shared variables -
among the ball and paddle classes - into a single space.

All necessary methods and variables have been initialised and are ready
for further coding.

This includes the position of the objects, their relative velocity and,
physical attributes (color & size) and their starting positions.