Experiment_3 Branch Commit
Big Update:

Have created two Enums - Player and Gamestate - to handle the various
players and game states.

Have created and added the Ball and Paddle classes - each has been
intitialised within the PongPanel class.

Added all playing objects into the PongPanel class - these include, the
drawing of the painted dotted line in the centre of the screen, the two
paddles - in the correct position and the ball - also in the correct
position - All this has been tested and all graphical elements are
working as of this update.

Added a number of clarifying comments on many of the methods and
variables within the project to make code more readable. - Hopefully
they make sense to someone other than just me

Nest Task:

Is to add movement to the objects and to allow the paddles to respond to
user input